Use a flexible attribute for mcrouter CLI options
Using the `node['mcrouter']['cli_opts']` attribute and `magic` cookbook’s
`shell_opts` method allows us to leave the usage of CLI options open-ended.

This results in a much more flexible cookbook: changes to CLI opts for
mcrouter won't necessitate trying to keep up in this cookbook.

Some options are necessary, and as such are set as defaults here:

* `port` & `config-file` are required by mcrouter
* `log-path` is needed to have any log output at all; e.g. for debugging
* `pid-file` is needed for Upstart to properly manage mcrouter

Others are set to their defaults (as of this writing) to ensure necessary directories are created:

* `async-dir`
* `stats-root`

Where possible, directory and file names are taken from attributes.