Refactor tests into Inspec to work with dokken
For some reason, kitchen-dokken does not load Serverspec tests into the
correct location.

It expects them to be at /tmp/verifier/suites/serverspec/\*\*/\*_spec.rb

But kitchen-dokken hooks them up to live at: /opt/verifier

Which is a Docker volume mapped to
~/.dokken/verifier_sandbox/suite-sandbox

Rather than wrestle with this, it’s known that Inspec tests place nicely
with dokken, so we refactor to get the same functionality with a
different tool.